// src/utils/var.d.ts
export declare enum BTNTYPE {
    QUOTE = "quote",
    H1 = "h1",
    H2 = "h2",
    H3 = "h3",
    BOLD = "bold",
    ITALIC = "italic",
    DELETE = "delete",
    UNDERLINE = "underLine",
    INLINECODE = "inlineCode",
    CODEBLOCK = "codeBlock",
    LINK = "link",
    NEWLINE = "newLine",
    ORDEREDLIST = "orderedList",
    UNORDERLIST = "unorderedList",
    IMAGE = "image",
    TABLE = "table",
    PREVIEW_MODE = "preview",
    EDIT_MODE = "edit",
    SYNC_MODE = "sync",
    MAXIMAZE = "maximize",
    MINIMAZE = "minimize"
}
export declare const InlineBtn: any;
export declare let left: BTNTYPE[];
export declare let right: BTNTYPE[];